server/requirements.txt
fastapi~=0.110.1
uvicorn~=0.29.0
pydantic-settings~=2.2.1
requests~=2.31.0
python-keycloak~=4.2.0
jwcrypto~=1.5.6

